requirements.txt
accelerate==1.6.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.2
cppimport==22.8.2
cupy-cuda12x==13.4.1
cycler==0.12.1
datasets==3.6.0
diffusers==0.33.1
dill==0.3.8
fastrlock==0.8.3
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.58.0
frozenlist==1.6.0
fsspec==2025.3.0
huggingface-hub==0.31.1
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.6
joblib==1.5.0
kiwisolver==1.4.8
lap==0.5.12
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.4.3
multiprocess==0.70.16
netifaces==0.11.0
networkx==3.4.2
ninja==1.11.1.4
numpy==2.2.5
nvidia-cuda-runtime-cu12==12.9.37
nvidia-ml-py==12.575.51
nvidia-modelopt==0.27.0
nvidia-modelopt-core==0.27.0
onnx==1.17.0
onnx_graphsurgeon==0.5.8
onnxconverter-common==1.14.0
onnxruntime-directml==1.22.0
onnxruntime-gpu==1.22.0
onnxslim==0.1.52
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
packaging==25.0
pandas==2.2.3
peft==0.15.2
pillow==11.2.1
platformdirs==4.3.8
propcache==0.3.1
protobuf==3.20.2
psutil==7.0.0
PuLP==3.1.1
py-cpuinfo==9.0.0
pyarrow==20.0.0
pybind11==2.13.6
pycuda==2025.1
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
pynvml==12.0.0
pyparsing==3.2.3
pyreadline3==3.5.4
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytools==2025.1.4
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
regex==2024.11.6
requests==2.32.3
rich==14.0.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
setuptools==80.4.0
shapely==2.1.0
siphash24==1.7
six==1.17.0
sympy==1.14.0
tensorrt==10.9.0.34
tensorrt_cu12==10.9.0.34
tensorrt_cu12_bindings==10.9.0.34
tensorrt_cu12_libs==10.9.0.34
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.0+cu128
torch-geometric==2.6.1
torch_tensorrt==2.7.0+cu128
torchaudio==2.7.0+cu128
torchprofile==0.0.4
torchvision==0.22.0+cu128
tqdm==4.67.1
transformers==4.51.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
ultralytics==8.3.132
ultralytics-thop==2.0.14
urllib3==2.4.0
xxhash==3.5.0
yarl==1.20.0
zipp==3.21.0
